All CRUD operations are completed in Role screen. Only minor checkbox related changes are pending

+++ RMDRMC.DataRepository.Sql/Interfaces/IRoleRepository.cs
@@ -19,7 +19,7 @@ namespace RMDRMC.DataRepository.Sql.Interfaces
 
         Roles GetRolesByName(string roleName);
 
-        bool DeleteRole(int roleID);
+        bool DeleteRole(long roleID);
        
     }
 }
